office/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
colorama==0.4.6
cryptography==39.0.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
filelock==3.9.0
future==0.18.3
gevent==22.10.2
greenlet==2.0.1
idna==3.4
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.3
tldextract==3.4.0
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.14
zope.event==4.6
zope.interface==5.5.2
